cleaned more files
