--- src/components/ListDisplay.jsx
@@ -1,9 +1,0 @@
-
-import  ItemDetails  from "./ItemDetails"
-import ListHead from "./ListHead"
-
-const ListDisplay = () => {
-    return (<div className="empty">No oustanding task!</div>)
-}
- 
-export default ListDisplay;
