--- src/components/ListHead.jsx
@@ -1,18 +1,0 @@
-import React from 'react';
-
-const ListHead = () => {
-
-
-    return (
-        <li>
-            <p>Complete</p>
-            <p>Importance</p>
-            <p>To do Item</p>
-        
-        </li>
-      );
-}
- 
-export default ListHead;
-
-
